+++ app/__init__.py
@@ -1,10 +1,8 @@
 from flask_login import LoginManager
 
-from app.db import Session
-from app.user.models import User
-from app.user.forms import LoginForm
-from app.user.views import blueprint as user_blueprint
-from app.edit.edit import edit
+from db import Session
+from user.views import blueprint as user_blueprint
+from edit.edit import edit
 from flask import Flask, abort, render_template
 from flask import session
 import os
@@ -20,6 +18,8 @@ def create_app():
     login_manager = LoginManager()
     login_manager.init_app(app)
     login_manager.login_view = 'login'
+
+    from models import User
 
     @login_manager.user_loader
     def user_loader(user_id) -> User:
