+++ app/models.py
@@ -1,11 +1,30 @@
 from datetime import datetime
 
+from flask_login import UserMixin
 from sqlalchemy import ForeignKey
-from sqlalchemy.orm import declarative_base, mapped_column, Mapped
+from sqlalchemy.orm import mapped_column, Mapped
+from werkzeug.security import generate_password_hash, check_password_hash
 
-from app.db import engine
+from db import Base, engine
 
-Base = declarative_base()
+
+class User(Base, UserMixin):
+    """Пользователи"""
+
+    __tablename__ = "users"
+    id: Mapped[int] = mapped_column(primary_key=True)
+    login: Mapped[str]
+    password: Mapped[str]
+    email: Mapped[str] = mapped_column(unique=True)
+
+    def __repr__(self):
+        return f'User login = {self.login}, email = {self.email}'
+
+    def set_password(self, password):
+        self.password = generate_password_hash(password)
+
+    def check_password(self, password):
+        return check_password_hash(self.password, password)
 
 
 class Account(Base):
