+++ app/test/test_database.py
@@ -1,18 +1,17 @@
-from flask import current_app
 from sqlalchemy.orm.session import Session
 import pytest
 from sqlalchemy import Engine, create_engine, select
 from sqlalchemy.orm import sessionmaker
 
-from app.crud import create_user, edit_transaction, edit_account
+from ..crud import create_user, edit_transaction, edit_account
 
-from app.models import Base
-from app.user.models import User
+from ..models import Base, User
 
 
 @pytest.fixture()
 def engine():
-    engine = create_engine(current_app.config.DB_URL, echo=True)
+    from ..config import DB_URL
+    engine = create_engine(DB_URL, echo=True)
     Base.metadata.create_all(engine)
     try:
         yield engine
@@ -53,7 +52,7 @@ def test_create_user(db_session: Session):
     user = db_session.execute(select(User).filter(User.login == login)).scalar()
     assert user is not None, "Пользователь не был создан"
     assert user.login == login, "Неправильный логин пользователя"
-    assert user.password == password, "Неправильный пароль пользователя"
+    assert user.check_password(password), "Неправильный пароль пользователя"
     assert user.email == email, "Неправильный email пользователя"
