--- requirements.txt
@@ -1,28 +1,0 @@
-black==23.12.1
-blinker==1.7.0
-click==8.1.7
-flake8==6.1.0
-Flask==3.0.0
-Flask-Login==0.6.3
-Flask-WTF==1.2.1
-greenlet==3.0.3
-iniconfig==2.0.0
-isort==5.13.2
-itsdangerous==2.1.2
-Jinja2==3.1.2
-MarkupSafe==2.1.3
-mccabe==0.7.0
-mypy-extensions==1.0.0
-packaging==23.2
-pathspec==0.12.1
-platformdirs==4.1.0
-pluggy==1.3.0
-psycopg2-binary==2.9.9
-pycodestyle==2.11.1
-pyflakes==3.1.0
-pytest==7.4.4
-python-dotenv==1.0.0
-SQLAlchemy==2.0.24
-typing_extensions==4.9.0
-Werkzeug==3.0.1
-WTForms==3.1.1
