+++ requirements/dev.txt
@@ -1,0 +1,7 @@
+-r prod.txt
+faker
+matplotlib
+black
+flake8
+isort
+colorama
