+++ requirements/prod.txt
@@ -1,0 +1,4 @@
+Flask>=3.0.0
+SQLAlchemy>=2.0.25
+WTForms>=3.1.1
+Flask-WTF>=1.2.1
