--- setup.cfg
@@ -1,8 +1,0 @@
-[tool.black]
-multi_line_output = 3
-include_trailing_comma = true
-line_length = 120
-ensure_newline_before_comments = true
-target-version = ['py311','py312']
-exclude = ".*,*/migrations/*,pycache,.venv/*"
-skip-string-normalization = true
