--- tox.ini
@@ -1,10 +1,0 @@
-[flake8]
-exclude = .*, venv
-accept-encodings = "utf-8"
-max-line-length = 120
-max-complexity = 6
-inline-quotes = "'"
-multiline-quotes = "'''"
-ignore=E203, E501, E704
-color=always
-
